+++ src/components/call/CallPage.tsx
@@ -1,13 +1,16 @@
 import * as React from 'react';
 import i18n from '../../services/i18n';
-import { LocationState } from '../../redux/location/reducer';
 import { RouteComponentProps } from 'react-router-dom';
 import { CallTranslatable } from './index';
-import { Layout } from '../shared';
+import { LayoutContainer } from '../layout';
 import { Issue } from '../../common/model';
 import { CallState, OutcomeData } from '../../redux/callState';
 
 /*
+  This is the top level View component in the CallPage Component Hierarchy.  It is the 
+    child of the Redux container.  Therefore, its "Props" property must match the
+    merged props that were provided to the connect() function in the "HomePageContainer".
+
     Note the "{id: string}" added as a generic type parameter to RouteComponentProps.
     If you look at the Type Definition F12(VSCode) for RouteComponentProps, you'll see this:
 
@@ -38,17 +41,25 @@ interface Props extends RouteProps {
   readonly onSubmitOutcome: (data: OutcomeData) => Function;
   readonly onSelectIssue: (issueId: string) => Function;
   readonly onGetIssuesIfNeeded: () => Function;
-
-  // location widget related
-  readonly locationState: LocationState;
-  readonly setLocation: (location: string) => void;
-  readonly clearLocation: () => void;
 }
 
 export interface State {
   currentIssue: Issue;
   callState: CallState;
 }
+
+/*
+  This is a StatelessComponent meaning that it is just a function. The props are passed in as
+  a property.  More complicated components will be instantiated as a class and will often
+  have "local" state.  Props for them will be an instance property.
+
+  Notice that we are just passing all of the props that we pull off the Redux Store through
+  this component to child components
+
+  When the props.onSelectIssue function is called by some component that has access to it
+  down this component hierarchy, it will simply be passed up this tree and end up calling the 
+  dispatch method on the store corresponding to that method(as defined in the top-level redux container). 
+*/
 
 class CallPage extends React.Component<Props, State> {
   constructor(props: Props) {
@@ -73,7 +84,7 @@ class CallPage extends React.Component<Props, State> {
     // Here we set it on the redux store(note that if we've already set it in local state, in this component)
     // we don't want to set it on the redux store again because that will cause a re-render loop.
     if (!this.props.callState.currentIssueId && newProps.currentIssue) {
-        this.props.onSelectIssue(newProps.currentIssue.id);
+      this.props.onSelectIssue(newProps.currentIssue.id);
     }
   }
 
@@ -93,15 +104,7 @@ class CallPage extends React.Component<Props, State> {
 
   render() {
     return (
-      <Layout
-        issues={this.props.issues}
-        completedIssueIds={this.props.callState.completedIssueIds}
-        currentIssue={this.props.currentIssue}
-        onSelectIssue={this.props.onSelectIssue}
-        locationState={this.props.locationState}
-        setLocation={this.props.setLocation}
-        clearLocation={this.props.clearLocation}
-      >
+      <LayoutContainer >
         {this.props.currentIssue &&
           <CallTranslatable
             issue={this.props.currentIssue}
@@ -109,7 +112,7 @@ class CallPage extends React.Component<Props, State> {
             onSubmitOutcome={this.props.onSubmitOutcome}
             t={i18n.t}
           />}
-      </Layout>
+      </LayoutContainer>
     );
   }
 }
