+++ src/components/done/DonePageContainer.tsx
@@ -1,8 +1,6 @@
 import { connect, Dispatch } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { ApplicationState } from '../../redux/root';
-import { newLocationLookup, clearAddress } from '../../redux/location';
-import { LocationState } from '../../redux/location/reducer';
 import { DonePage } from './index';
 import { Issue } from '../../common/model';
 import { getIssuesIfNeeded } from '../../redux/remoteData';
@@ -14,17 +12,13 @@ interface OwnProps extends RouteComponentProps<{ id: string }> { }
 
 interface StateProps {
   readonly issues: Issue[];
-  readonly completedIssueIds: string[];
   readonly currentIssue?: Issue;
   readonly totalCount: number;
-  readonly locationState: LocationState;
 }
 
 interface DispatchProps {
   readonly onSelectIssue: (issueId: string) => void;
   readonly onGetIssuesIfNeeded: () => void;
-  readonly setLocation: (location: string) => void;
-  readonly clearLocation: () => void;
 }
 
 const mapStateToProps = (state: ApplicationState, ownProps: OwnProps): StateProps => {
@@ -35,10 +29,8 @@ const mapStateToProps = (state: ApplicationState, ownProps: OwnProps): StateProp
 
   return {
     issues: state.remoteDataState.issues,
-    completedIssueIds: state.callState.completedIssueIds,
     currentIssue: currentIssue,
     totalCount: state.remoteDataState.callTotal,
-    locationState: state.locationState,
   };
 };
 
@@ -47,8 +39,6 @@ const mapDispatchToProps = (dispatch: Dispatch<ApplicationState>): DispatchProps
     {
       onSelectIssue: selectIssueActionCreator,
       onGetIssuesIfNeeded: getIssuesIfNeeded,
-      setLocation: newLocationLookup,
-      clearLocation: clearAddress,
     },
     dispatch);
 };
