+++ src/components/faq/FaqPage.tsx
@@ -1,32 +1,11 @@
 import * as React from 'react';
-import { LocationState } from '../../redux/location/reducer';
-import { RouteComponentProps } from 'react-router-dom';
 import { Faq } from './index';
-import { Layout } from '../shared/index';
-import { Issue } from '../../common/model';
+import { LayoutContainer } from '../layout';
 
-interface Props extends RouteComponentProps<{ id: string }> {
-  readonly issues: Issue[];
-  readonly completedIssueIds: string[];
-  readonly onSelectIssue: (issueId: string) => Function;
-
-  // location widget related
-  readonly locationState: LocationState;
-  readonly setLocation: (location: string) => void;
-  readonly clearLocation: () => void;
-}
-
-const FaqPage: React.StatelessComponent<Props> = (props: Props) => (
-  <Layout
-    issues={props.issues}
-    completedIssueIds={props.completedIssueIds}
-    onSelectIssue={props.onSelectIssue}
-    locationState={props.locationState}
-    setLocation={props.setLocation}
-    clearLocation={props.clearLocation}
-  >
+const FaqPage: React.StatelessComponent<{}> = (props: {}) => (
+  <LayoutContainer>
     <Faq />
-  </Layout>
+  </LayoutContainer>
 );
 
 export default FaqPage;
