+++ src/components/layout/index.ts
@@ -1,0 +1,9 @@
+import Layout from './Layout';
+import LayoutContainer from './LayoutContainer';
+import Navigation from './Navigation';
+import Sidebar from './Sidebar';
+import Footer from './Footer';
+
+export {
+  Layout, Navigation, Sidebar, LayoutContainer, Footer
+};
