+++ src/components/shared/index.ts
@@ -1,12 +1,6 @@
-import Layout from './Layout';
-import Navigation from './Navigation';
 import Promotion from './Promotion';
-import Sidebar from './Sidebar';
 import CallCount from './CallCount';
-import Footer from './Footer';
 
 export {
-  Layout, Navigation, Promotion,
-  Sidebar, CallCount,
-  Footer
+  Promotion, CallCount
 };
